service: move accountid, credential, locker to internal
