+++ service/controller/resource/ipam/create.go
@@ -9,7 +9,7 @@ import (
 	"github.com/giantswarm/microerror"
 	"k8s.io/apimachinery/pkg/api/meta"
 
-	"github.com/giantswarm/aws-operator/service/locker"
+	"github.com/giantswarm/aws-operator/service/internal/locker"
 )
 
 // EnsureCreated allocates tenant cluster network segments. It gathers existing
